+++ src/watcher.ts
@@ -15,28 +15,36 @@ export class Watcher {
   readonly events: Watcher.Events
   readonly finder: Watcher.Options['finder']
 
-  private constructor(watchPath: string | string[], opts: Watcher.Options) {
+  private constructor(watchPath: string[], opts: Watcher.Options) {
     this.converter = opts.converter
     this.finder = opts.finder
     this.events = opts.events
 
-    this.chokidar = chokidar.watch(watchPath, { ignoreInitial: true })
-    this.chokidar
+    this.chokidar = chokidar
+      .watch(watchPath, { disableGlobbing: true, ignoreInitial: true })
       .on('change', f => this.convert(f))
       .on('add', f => this.convert(f))
+      .on('unlink', f => this.delete(f))
+
+    this.converter.options.themeSet.onThemeUpdated = f => this.convert(f)
 
     notifier.start()
 
     cli.info(chalk.green('[Watch mode] Start watching...'))
   }
 
-  private async convert(fn) {
-    const files = (await this.finder()).filter(
-      f => path.resolve(f.path) === path.resolve(fn)
+  private async convert(filename: string) {
+    const resolvedFn = path.resolve(filename)
+    const mdFiles = (await this.finder()).filter(
+      f => path.resolve(f.path) === resolvedFn
+    )
+    const cssFile = (await this.converter.options.themeSet.findPath()).find(
+      f => path.resolve(f) === resolvedFn
     )
 
+    // Update markdown
     try {
-      await this.converter.convertFiles(files, ret => {
+      await this.converter.convertFiles(mdFiles, ret => {
         this.events.onConverted(ret)
 
         if (ret.file.type === FileType.File)
@@ -45,9 +53,20 @@ export class Watcher {
     } catch (e) {
       this.events.onError(e)
     }
+
+    // Reload Theme CSS
+    if (cssFile !== undefined) this.converter.options.themeSet.load(resolvedFn)
   }
 
-  static watch(watchPath: string | string[], opts: Watcher.Options) {
+  private delete(filename: string) {
+    const fn = path.resolve(filename)
+    const { themeSet } = this.converter.options
+
+    themeSet.unobserve(fn)
+    themeSet.themes.delete(fn)
+  }
+
+  static watch(watchPath: string[], opts: Watcher.Options) {
     return new Watcher(watchPath, opts)
   }
 }