+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add `--theme-set` option to use additional theme CSS files ([#21](https://github.com/marp-team/marp-cli/pull/21))
+- Support auto reloading of additional theme CSS in watch mode ([#22](https://github.com/marp-team/marp-cli/pull/22))
 
 ### Changed